backend/requirements.txt
sanic
groq
python-dotenv